handled case where on paginating sort type would get messed up.
RCA: When paginated, we would call getFeed() method again with sortType as empty if we are on frontpage. The same would be set in the UI on api response.
This was not happening for the first time because we had default sort type as hot. Other wise this would have shown up in the first load as well.

Fix: Now we have a method in view model which checks if subreddit is of front page and if sort type is empty then it returns sort type as best, otherwise return whichever sort type is present. The result of this method is being set as sort type for UI state before returning the state to UI